pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfb-schedule-tracker"
version = "0.1.0"
description = "College Football Schedule Tracker with Calendar Integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jesse Oseafiana"}
]
keywords = ["college-football", "schedule", "calendar", "cfbd", "sports"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
]

dependencies = [
    "cfbd>=5.0.0",
    "icalendar>=5.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=1.10.0,<2.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
gcal = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.scripts]
cfb-tracker = "cfb_tracker.cli:main"

[project.urls]
Homepage = "https://github.com/jesseoseafiana/cfb-schedule-tracker"
Repository = "https://github.com/jesseoseafiana/cfb-schedule-tracker"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
